Wrap public methods in locks to ensure atomicity
